refactor(services): extract decodeCredential to common utility
Move decode logic from VCKit adapter to verifiable-credential/common
since JWT decoding is adapter-agnostic. Remove decode from the
IVerifiableCredentialService interface, fix maskInstanceConfig logger
type to accept pino loggers, make credential tenantId required in
list queries, and include decryptionKey in credential GET response.
